Rewrite graph processing
This rewrites how we build up the dependency graph, to rely on the
various metadata found in the exported equinox source files.  We rely
slightly less on manual parsing of the lexed text in places now.

The output is so far unchanged, apart from some differences to the
capitalisation of node labels, caused by the switch from using partial
filenames to using information from within the source file itself.