Fix merged node bug
We were generating node identifiers based purely on the module name, and
not on the type.  This means that nodes were merged for example when a
method and a form of the same name both existed.  We now use the type as
a part of the input for generating the identifier.