Switch to tracking calls by `node`
This makes a subtle but (especially in future) important switch from
tracking by "module" which roughly means "source file" to `node`
instead.  One important example of the difference, is that public
procedures are nodes (i.e, a callable unit of executable) whereas they
are not a module.

This also involved removing processing of calls from "procedure modules"
to methods.  There are very few of these anyway, but it doesn't make
much sense, and the eventual goal is to replace this with calls from the
procedures themselves instead.  Meanwhile, they are just missing for
now.