Add support for special cases
There are some special cases where we might want to make certain nodes
in particular ways that cannot be automatically derived.  This
introduces the specialJson parameter for this.

The only supported feature right now is to list "system procedures"
which we will mark as used, even if there is no explicit use in the
analysed source code.